refactor: move public wallet setting to shared
