+++ packages/plugins/Web3Profile/src/SNSAdaptor/index.tsx
@@ -1,5 +1,5 @@
 import { Plugin, PluginID } from '@masknet/plugin-infra'
-import { ApplicationEntry } from '@masknet/shared'
+import { ApplicationEntry, PublicWalletSetting } from '@masknet/shared'
 import { Icons } from '@masknet/icons'
 import { base } from '../base.js'
 import { Web3ProfileDialog } from './components/Web3ProfileDialog.js'
@@ -8,7 +8,6 @@ import { PluginI18NFieldRender } from '@masknet/plugin-infra/content-script'
 import { Trans } from 'react-i18next'
 import { CrossIsolationMessages } from '@masknet/shared-base'
 import { useEffect } from 'react'
-import { PublicWalletSetting } from './components/PublicWalletSetting'
 
 const sns: Plugin.SNSAdaptor.Definition = {
     ...base,
