+++ packages/shared/src/UI/components/index.ts
@@ -35,3 +35,4 @@ export * from './FootnoteMenu/index.js'
 export * from './MaskPluginWrapper/index.js'
 export * from './MaskIcon/index.js'
 export * from './SourceProviderIcon/index.js'
+export * from './PublicWalletSetting'
