+++ packages/shared/src/locales/en-US.json
@@ -109,5 +109,14 @@
     "plugin_card_frame_default_title": "Web3 Profile",
     "plugin_card_frame_default_provided_by": "Powered by",
     "plugin_card_frame_default_provided_by_value": "Mask Network",
-    "plugin_card_frame_loading": "Loading...."
+    "plugin_card_frame_loading": "Loading....",
+    "save": "Save",
+    "cancel": "Cancel",
+    "add_wallet": "Add Wallet",
+    "wallet_setting_hint": "Toggle the button to manage wallet display settings.",
+    "add_wallet_tips": "Please add wallets to receive tips",
+    "save_successfully": "Save successfully",
+    "save_failed": "Save failed",
+    "wallet_set_up_successfully": "Web3 Profile Wallet set up successfully.",
+    "wallet_set_up_failed": "Web3 Profile Wallet setup failed, please try again."
 }
